Widen fork outcome indexes to uint256 and harden forgeries
- Convert security-fork outcome paths from `uint8` to `uint256` across Solidity, TS ABI typings, and UI contract calls
- Add constructor/runtime guards in `Zoltar` to fail fast on missing genesis REP code/supply and uninitialized universe forks
- Add oracle coordinator recovery path for failed settlement callbacks with new `recoverSettledPendingReport` flow
- Extend regression/security tests for external scalar forks, pending oracle recovery, and fork-data storage/typing regressions
- Update simulator/bootstrap typing/constants to keep synthetic fork flows consistent

+++ solidity/contracts/peripherals/SecurityPoolForkerTypes.sol
@@ -19,7 +19,7 @@ struct SecurityPoolForkerForkData {
 	bool ownFork;
 	uint256 vaultRepAtFork;
 	bool unresolvedEscalationAtFork;
-	uint8 outcomeIndex;
+	uint256 outcomeIndex;
 	uint256 ownForkCollateralAtFork;
 	uint256 ownForkMigratedRepCollateralized;
 	uint256 ownForkCollateralTransferred;
